Relocate MySQL classes to their own package
