+++ common/src/main/java/net/draycia/carbon/common/CarbonCommonModule.java
@@ -42,7 +42,7 @@ import net.draycia.carbon.common.messages.placeholders.StringPlaceholderResolver
 import net.draycia.carbon.common.messages.placeholders.UUIDPlaceholderResolver;
 import net.draycia.carbon.common.users.CarbonPlayerCommon;
 import net.draycia.carbon.common.users.JSONUserManager;
-import net.draycia.carbon.common.users.db.MySQLUserManager;
+import net.draycia.carbon.common.users.db.mysql.MySQLUserManager;
 import net.kyori.adventure.audience.Audience;
 import net.kyori.adventure.key.Key;
 import net.kyori.adventure.text.Component;
